Rename demo_app to demoapp-go
