+++ .circleci/continue-workflows.yml
@@ -42,7 +42,7 @@ parameters:
   updated-aperture-charts:
     type: boolean
     default: false
-  updated-demo-app:
+  updated-demoapp-go:
     type: boolean
     default: false
 
@@ -508,7 +508,7 @@ workflows:
           use-docker-layer-caching: true
 
   demo-app:
-    when: << pipeline.parameters.updated-demo-app >>
+    when: << pipeline.parameters.updated-demoapp-go >>
     jobs:
       - build-push-add-tag:
           name: image-demo-app
@@ -516,7 +516,7 @@ workflows:
           gcr-registry: gcr.io/devel-309501
           gcr-image: cf-fn/demo-app
           docker-context: .
-          dockerfile: tools/demo_app/Dockerfile
+          dockerfile: tools/demoapp-go/Dockerfile
           use-docker-layer-caching: true
           push-to-dockerhub: false
