+++ tools/demoapp-go/Dockerfile
@@ -9,10 +9,10 @@ COPY --link . .
 RUN --mount=type=cache,target=/go/pkg/ \
     --mount=type=cache,target=/root/.cache/go-build/ <<-EOF
     go mod download
-    CGO_ENABLED=0 go build -ldflags="-w -s" -o /demo_app ./tools/demo_app/cmd/demo_app
+    CGO_ENABLED=0 go build -ldflags="-w -s" -o /demoapp-go ./tools/demoapp-go
 EOF
 
 # Final image
 FROM alpine:3.15.0
-COPY --link --from=builder /demo_app /demo_app
-ENTRYPOINT ["/demo_app"]
+COPY --link --from=builder /demoapp-go /demoapp-go
+ENTRYPOINT ["/demoapp-go"]
