+++ tools/demoapp-go/main.go
@@ -14,7 +14,7 @@ import (
 	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
 
 	"github.com/fluxninja/aperture/pkg/log"
-	"github.com/fluxninja/aperture/tools/demo_app/app"
+	"github.com/fluxninja/aperture/tools/demoapp-go/app"
 )
 
 func main() {
